+++ st-exerciseviewer/src/main/kotlin/de/saring/exerciseviewer/gui/panels/LapPanelController.kt
@@ -106,11 +106,11 @@ class LapPanelController(
         }
         tcEndSpeed.cellFactory = FormattedNumberCellFactory {
             if (it == null) null
-            else context.formatUtils.speedToString(it.toFloat(), 2, document.speedMode)
+            else context.formatUtils.speedToString(it.toDouble(), 2, document.speedMode)
         }
         tcAvgSpeed.cellFactory = FormattedNumberCellFactory {
             if (it == null) null
-            else context.formatUtils.speedToString(it.toFloat(), 2, document.speedMode)
+            else context.formatUtils.speedToString(it.toDouble(), 2, document.speedMode)
         }
         tcDistance.cellFactory = FormattedNumberCellFactory {
             if (it == null) null
