+++ st-exerciseviewer/src/main/kotlin/de/saring/exerciseviewer/gui/panels/OptionalPanelController.kt
@@ -70,9 +70,9 @@ import javafx.scene.control.Label
 
         // fill speed data
         exercise.speed?.let {
-            laSpeedAvgValue.text = formatUtils.speedToString(it.speedAvg, 2, document.speedMode)
-            laSpeedMaxValue.text = formatUtils.speedToString(it.speedMax, 2, document.speedMode)
-            laDistanceValue.text = formatUtils.distanceToString((it.distance / 1000f).toDouble(), 2)
+            laSpeedAvgValue.text = formatUtils.speedToString(it.speedAvg.toDouble(), 2, document.speedMode)
+            laSpeedMaxValue.text = formatUtils.speedToString(it.speedMax.toDouble(), 2, document.speedMode)
+            laDistanceValue.text = formatUtils.distanceToString((it.distance / 1000.0), 2)
         }
 
         exercise.recordingMode.bikeNumber?.let {
